Type-disagreement repair 4: 89 shadow entries closed, zephyr.bin BYTE-IDENTICAL
89 shadow-class entries closed (236 -> 147 conflicting; 525 -> 330 declaration
sites) with the linked zephyr.bin BYTE-IDENTICAL to the baseline (cmp exit 0) --
a stronger gate than pass 3, which could only compare the app_update.bin payload.

BASELINE MOVED UNDER US
DEFECTS.json was regenerated by the concurrent refactor agent: 934, not 949.
Stage 03's agreement test at HEAD: 715, not 721. Both reported; the two ledgers
are not term-comparable.

TASK 1 -- SHADOW CLASS SCALED, WITH CONTROLS ON EVERY STEP
Pass 3's two stated blockers removed, both controlled:
- SDK parameter types put in scope (hdrprobe.py, with a NO-HEADER POSITIVE
  CONTROL that fails for every type): 188 compile errors -> 76.
- GCC reports `previous *definition*` for a static inline: harvest 82/109 ->
  107/109 symbols.
- A POSITIVE CONTROL ON THE HARVEST DROPPED 3 (memcpy, memmove, strcpy -- the
  SDK's own prototype does NOT clear their conflict).
- The all-at-once applier had to become INCREMENTAL per (file, symbol): the bigger
  candidate set made it WORSE (76 -> 140 error files) until one symbol at a time
  was gated.
Zero flash cost holds at scale: 188 edited files, 177 GATE A, 11 not build inputs
(none #include'd), image byte-identical.

CALIBRATION BUG 8 (inherited from pass 3)
The shadow probe scores ANY compiler error as a type conflict. When the SDK
defines the name as a FUNCTION-LIKE MACRO, the declaration text macro-expands into
a SYNTAX error. 22 of the 147 residual entries are that artifact and CAN NEVER BE
CLOSED BY RESPELLING; 4 more are a probe artifact. Only 121 are genuine type
conflicts.

THE RESIDUE IS ONE SHAPE, AND IT WAS MEASURED
Failures are dominated by k_timeout_t -- an 8-byte struct in an aligned register
pair. The shipped disassembly of att_req_sent proves our 4-unsigned-int spelling
fills EXACTLY THE REGISTERS THE SDK PROTOTYPE FILLS: ABI-correct, source-wrong.
Repaired via compound literals and measured: 42 objects, +8 B total, 40 exactly
size-neutral at nm -S function granularity, difference is INSTRUCTION SCHEDULING
ONLY. QUARANTINED, NOT LANDED -- but it dominates pass 3's constant-argument batch
(+66 B, 8 size-neutral, and that one changes what the callee receives).
Recommended as the first batch for a Renode-owning pass.

A REGRESSION CAUSED AND REVERTED
Canonicalisation of shadowed symbols improved Stage 03's test 715 -> 697 but
REGRESSED the shadow class 147 -> 155, because 15 of the 18 "fixes" were on the
class where that test is vacuous. Both rounds reverted -- a plain `rsync -a`
missed one file, caught by re-counting 157 != 156 -- and re-run with shadow
symbols excluded.

FOUR-WAY SPLIT (934 entries)
FIXED 5 + 89 shadow; CONFIRMED REAL 149; FALSE ALARM 172 (18.4 %); UNSETTLED 377.
The 231 the script prints as FIXED for the shadow class is the vacuous artifact
and is NOT claimed. Agreement test 715 -> 698, of which 16 of 17 are on the
vacuously-scored class -- only 1 real non-shadow closure.

DEAD FUNCTIONS -- QUESTION ANSWERED
Stage 07 built read-only; 118 -Wunused-function. Name-based intersection gives 0,
address-based gives 16 -- the gap is the alias confound, including a THIRD
provenance form pass 3 did not name (the inline /*=0xADDR*/ marker). After
resolving all three: 14 survive, and only 2 have a shipped caller that is a file
of ours. None is one of pass 3's 12 refined lost callees. A lead; nothing changed.

ACCEPTANCE
nm -u 0; 0 dup globals; app pins 627/0/0; net pins 0/0 on the FROZEN 225,581 B
image (not rebuilt); check_thread_create_stack_args 10/10; verify_data 995/995;
FLASH 956,480 B delta 0; RAM delta 0; 3,482/3,484 objects identical
debug-stripped (the 2 are uname.c's build clock and an embedded __FILE__ path).
FRAMEBUFFERS NOT CAPTURED AND NOT CLAIMED -- Renode belonged to a concurrent
agent. The only claim is the inference from zephyr.bin byte-equality.

Detail: recon/analysis/type_disagreement_repair_4.md

+++ recon/named/adv_is_directed.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 /* readable reconstruction; identity: FUN_00055698 @ 0x00055698
  * public-name: adv_is_directed
  * durable-map: recon/catalogs/function_names_app.json
@@ -8,5 +9,5 @@
  * Raw backmap: FUN_00055698@0x00055698. */
 
 #include <stdint.h>
-extern int memcmp(const void*,const void*,uint32_t);
+extern int memcmp(const void *, const void *, size_t);
 int adv_is_directed(const void *advertisement){return memcmp((const uint8_t*)advertisement+9u,(const void*)0x000f2b3au,7u)!=0;}
